Use biblatex/biber instead of natbib/bibtex
As natbib and bibtex are not maintained and can't handle
UTF-8 input, the Tufte-LaTeX class should use biblatex and
biber instead.

Note, however, that this change breaks the vertical offset
argument of the `\cite` command due to how biblatex handles
that command. This may cause problems with citations running
off the bottom of the pages; but since pure-citation sidenotes
are likely to be relatively short, this seems like an acceptable
loss.

Ref #60.